Add docker basics and pymongo
Add the basics for a docker based deployment.  This should make it
easier to run everywhere although more tweaking may be required.

Add pymongo to allow the storage of documents in a document database.
This may be replaced by a relational database later as there is some
valuable relationships here.  Another option is to stage in the mongodb
and once all the data is scraped create a relational database for
analysis.